streamlit_requirements.txt
streamlit>=1.28.0
opencv-python>=4.5.0
numpy>=1.20.0
pandas>=1.3.0
plotly>=5.0.0
Pillow>=8.0.0
